Drop unused test images, constants and helpers
