--- test/test_images/helloworld/helloworld.yaml
@@ -1,40 +1,0 @@
-# Copyright 2018 The Knative Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: serving.knative.dev/v1
-kind: Configuration
-metadata:
-  name: configuration-example
-  namespace: default
-spec:
-  template:
-    spec:
-      containers:
-      - # This is the Go import path for the binary to containerize
-        # and substitute here.
-        image: ko://knative.dev/networking/test/test_images/helloworld
-        readinessProbe:
-          httpGet:
-            path: /
-          initialDelaySeconds: 3
----
-apiVersion: serving.knative.dev/v1
-kind: Route
-metadata:
-  name: route-example
-  namespace: default
-spec:
-  traffic:
-  - configurationName: configuration-example
-    percent: 100
