--- test/test_images/helloworld/service.yaml
@@ -1,10 +1,0 @@
-apiVersion: serving.knative.dev/v1
-kind: Service
-metadata:
-  name: helloworld-test-image
-  namespace: default
-spec:
-  template:
-    spec:
-      containers:
-      - image: ko://knative.dev/networking/test/test_images/helloworld
