+++ test/util.go
@@ -18,17 +18,23 @@ package test
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"net/http"
+	"testing"
 	"time"
 
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	pkgnet "knative.dev/pkg/network"
 	"knative.dev/pkg/signals"
-	"knative.dev/pkg/test/logging"
-	"knative.dev/pkg/test/spoof"
+	pkgTest "knative.dev/pkg/test"
+	"knative.dev/pkg/test/helpers"
+	"knative.dev/pkg/test/logstream"
+
+	// For our e2e testing, we want this linked first so that our
+	// systen namespace environment variable is defaulted prior to
+	// logstream initialization.
+	_ "knative.dev/networking/test/defaultsystem"
+
+	// Mysteriously required to support GCP auth (required by k8s libs). Apparently just importing it is enough. @_@ side effects @_@. https://github.com/kubernetes/client-go/issues/242
+	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 )
 
 const (
@@ -36,15 +42,26 @@ const (
 	PollInterval = 1 * time.Second
 	// PollTimeout is how long e2e tests will wait for resource updates when polling.
 	PollTimeout = 10 * time.Minute
-
-	// HelloVolumePath is the path to the test volume.
-	HelloVolumePath = "/hello/world"
-
-	caSecretNamespace = "cert-manager"
-	caSecretName      = "ca-key-pair"
 )
 
 // util.go provides shared utilities methods across knative serving test
+
+// Setup creates client to run Knative Service requests
+func Setup(t testing.TB) *Clients {
+	t.Helper()
+
+	cancel := logstream.Start(t)
+	t.Cleanup(cancel)
+
+	clients, err := NewClients(pkgTest.Flags.Kubeconfig, pkgTest.Flags.Cluster, ServingNamespace)
+	if err != nil {
+		t.Fatal("Couldn't initialize clients", "error", err.Error())
+	}
+	return clients
+}
+
+// ObjectNameForTest generates a random object name based on the test name.
+var ObjectNameForTest = helpers.ObjectNameForTest
 
 // ListenAndServeGracefully calls into ListenAndServeGracefullyWithPattern
 // by passing handler to handle requests for "/"
@@ -61,34 +78,4 @@ func ListenAndServeGracefullyWithHandler(addr string, handler http.Handler) {
 
 	<-signals.SetupSignalHandler()
 	server.Shutdown(context.Background())
-}
-
-// AddRootCAtoTransport returns TransportOption when HTTPS option is true. Otherwise it returns plain spoof.TransportOption.
-func AddRootCAtoTransport(ctx context.Context, logf logging.FormatLogger, clients *Clients, https bool) spoof.TransportOption {
-	if !https {
-		return func(transport *http.Transport) *http.Transport {
-			return transport
-		}
-	}
-	rootCAs, _ := x509.SystemCertPool()
-	if rootCAs == nil {
-		rootCAs = x509.NewCertPool()
-	}
-	if !rootCAs.AppendCertsFromPEM(PemDataFromSecret(ctx, logf, clients, caSecretNamespace, caSecretName)) {
-		logf("Failed to add the certificate to the root CA")
-	}
-	return func(transport *http.Transport) *http.Transport {
-		transport.TLSClientConfig = &tls.Config{RootCAs: rootCAs}
-		return transport
-	}
-}
-
-// PemDataFromSecret gets pem data from secret.
-func PemDataFromSecret(ctx context.Context, logf logging.FormatLogger, clients *Clients, ns, secretName string) []byte {
-	secret, err := clients.KubeClient.CoreV1().Secrets(ns).Get(ctx, secretName, metav1.GetOptions{})
-	if err != nil {
-		logf("Failed to get Secret %s: %v", secretName, err)
-		return []byte{}
-	}
-	return secret.Data[corev1.TLSCertKey]
 }
